# test: speed up tests