--- a/jest.config.js
+++ b/jest.config.js
@@ -3,4 +3,9 @@ module.exports = {
   preset: 'ts-jest',
   testEnvironment: 'node',
   modulePathIgnorePatterns: ['<rootDir>/test/'],
+  globals: {
+    'ts-jest': {
+      isolatedModules: true,
+    },
+  },
 };