## Before
```
Executed in   35.98 secs    fish           external
   usr time  336.41 secs  192.00 micros  336.41 secs
   sys time   48.36 secs  683.00 micros   48.36 secs
```

 ## After
```
Executed in    5.98 secs    fish           external
   usr time   21.76 secs  194.00 micros   21.76 secs
   sys time    5.31 secs  493.00 micros    5.31 secs
```